END: Add training and performance environment configuration (#1794)
This PR adds E2E configuration to target the training and performance environments. This was necessary to support the load of the training data, and I just went ahead and did performance at the same time. We should not target any E2E tests at the training environment right now - this is for future use (and biz sim use) only.