END-459 Silence Cypress error appearing in New Relic, update local config (#1802)
https://lwd.atlassian.net/browse/END-459

Cypress and New Relic don't play very well together - they produce a persistent runtime error:

The following error originated from your application code, not from Cypress. > Cannot set property 'status' of undefined When Cypress detects uncaught errors originating from your application it will automatically fail the current test. This behavior is configurable, and you can choose to turn this off by listening to the `uncaught:exception` event.
This error is being picked up repeatedly in New Relic browser, leading to a lot of noise for the Portal team. Unfortunately, we can't fix this error (it's deep in the internals of both Cypress and New Relic), but we can silence it from being reported to New Relic, which should cut the noise.

This PR also:

Removes cypress-unfetch and switches to the new route2 API. I made this change to try to resolve the error. It doesn't work, but it also doesn't break anything, so I don't see a reason to stick with cypress-unfetch.
Updates Cypress (again, done in an effort to resolve. Also doesn't work, but no reason not to update unless it breaks something).
Updates config.json to be appropriate for local environments.
This PR doesn't:

Fix the "Unhandled Runtime Error" message that pops up when running the portal locally. That's a NextJS issue I think we have to work with the portal team on to find another way around.

* END-459: Silence Cypress error with New Relic browser

* Update E2E configuration for local environment

* Remove cypress-unfetch

* Update typings for better task completion