API-833: refactor documents check (#1729)
🎫 Ticket
https://lwd.atlassian.net/browse/API-833

🛠 Changes
Refactor get_document_by_id to search through all documets.

ℹ️ Context for reviewers
The code before this change would throw an error if the first document in the documents list didn't match the sought-after document. This change allows the rest of the documents to be searched through before the No document found for ID error is raised.