END: Fix claim ID extraction in Cypress tests (#1826)
Last week we updated to use cy.route2 instead of cypress-unfetch to capture HTTP requests in Cypress. I made a mistake in the implementation - cy.route2 doesn't seem to reliably unserialize the response body from JSON (it's a string instead of an object). As a result, we were getting failures in reading the Fineos ID off of the response body. This corrects the issue.

* END: Fix claim ID extraction in Cypress tests

* Fix CS