API: Configure `stage` for the dev FINEOS IAM role (#1818)
It was originally thought that the IDT buckets would be in the pre-live
FINEOS environment, but they exist in the dev environment. The bucket
path was already updated[1] for this, but the IAM role was still
configured for the UAT/pre-live environment. Update that to use the dev
environment role.

[1] c95483aee61026fbf7919040b3ea2ca4bc3426b4