remove cloudtrail and associated buckets (#1834)
🎫 Ticket
https://lwd.atlassian.net/browse/API-820

🛠 Changes
removed the massgov-pfml-{env-name}-logging cloudtrails and their associated buckets.

ℹ️ Context for reviewers
Delete the S3 object log cloudtrails, since Smartronix is going to manage them instead

in addition, there is a 5 trail limit for Cloudtrail. Exceeding this limit causes some deploys to error out. Example
Essentially, the goal of this pr is to undo what was added here