CP-1255 Add employer benefits and other incomes API (#1648)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-1255

## 🛠 Changes

* Adds five new fields on the Application:
  * has_employer_benefits - boolean
  * has_other_incomes - boolean
  * other_incomes_awaiting_approval - boolean
  * employer_benefits - array of EmployerBenefits
  * other_incomes - array of OtherIncomes
* Adds models and tables for EmployerBenefit and OtherIncome types

## ℹ️ Context for reviewers

Note that this PR doesn't handle deleting existing employer benefits or other incomes. There doesn't seem to be a set way of doing that right now. [API-505](https://lwd.atlassian.net/browse/API-505) has some comments discussing ways to do this for leave periods which should probably be similar to reductions.

The API for previous leaves will be done as a separate ticket.