API: disable lookup table syncs in Lambda functions and optimize (#1912)
## 🎫 Ticket

None

## 🛠 Changes

- Add `sync_lookups` parameter to `massgov.pfml.db.init()` that can be used to skip synchronization of lookup tables.
- Set `sync_lookups=True` in the API server only.
- Add an optimization to `massgov.pfml.db.lookup.sync_to_database()` to reduce db queries.

## ℹ️ Context for reviewers

Synchronizing lookup tables takes somewhere between 0.7-1s according to logs. This cuts into the time allowed for Cognito lambdas. We don't need to do it every time - we can leave this task to the API server start.

The optimization was observed to cut about 0.4s from the largest table (LkCountry). It is less significant for small tables.

Reviewers: @kyeah @doshitan