API-133: Build API release image from scratch (#1533)
https://lwd.atlassian.net/browse/API-133

## 🛠 Changes

Builds the docker release image from `scratch` (an empty base image) instead of `python:3.8-slim`. This removes access to bash, sh, and many other basic development tools that are unnecessary when running the application in deployed environments.

## ℹ️ Context for reviewers

The following required files are included to allow the application to continue running:

- SSL/CA Certificates for making HTTPS requests
- python binaries and dependent libs

## 👀 Demo

Verified that the API still starts up locally:

```
make release-start-api
```

and that I can visit swagger.