EDM-72: Update employee/employer pool for test (#2065)
https://lwd.atlassian.net/browse/END-72

This PR updates the employee/employer pool that we're using for testing to reference the new pool of 10k dummy users. These users should have working Financial eligibility in both test and stage, and the FEINs contained should work in both test and stage for the leave admin self registration test.

Note: As of this writing, the data set hasn't been imported into stage yet - that should happen soon. We're waiting on a Fineos release.

* EDM-72: Update employee/employer pool for test

* Update config for employee pool as well

* Remove old test data