add email address to fineos contact details patch (#2137)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-1486

## 🛠 Changes

- Adds user's email address to FINEOS `/updateContactDetails` post

## ℹ️ Context for reviewers

In #1999 we established a pattern of storing the phoneNumber ID from FINEOS and sending it with every update call. Email address is the same situation, except we are not storing the FINEOS email address ID. However, the FINEOS API overwrites the email address with each request (see terminal snapshot in demo), so, though sending the ID is ideal, it's not strictly required. In the interest of time, i've used the simpler approach, but am open to making those changes now if reviewers prefer.

## 👀 Demo

### terminal 
<img width="943" alt="Screen Shot 2020-12-03 at 1 30 26 PM" src="https://user-images.githubusercontent.com/2473869/101090613-c5b9e800-356b-11eb-9deb-3560dcafe607.png">

### fineos
<img width="338" alt="Screen Shot 2020-12-03 at 1 07 16 PM" src="https://user-images.githubusercontent.com/2473869/101089706-64dde000-356a-11eb-98b7-3ccf1a07a169.png">


## 🌡 End to End testing

N/A

## 🧐 
@lawrencekang