API-991: Increase keepalive timeout on gunicorn workers to be higher than NLB (#2221)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-991

## 🛠 Changes

- Increase the keepalive from the default 2 seconds, up to 355 seconds.

## ℹ️ Context for reviewers

When dealing with a load balancer (NLB, ELB, etc.) it is recommended that you use a higher keepalive timeout on your application servers than the load balancer, see [Amazon Docs](http://docs.aws.amazon.com/ElasticLoadBalancing/latest/DeveloperGuide/ts-elb-error-message.html).

> Cause 2: Registered instances closing the connection to Elastic Load Balancing.
>
> Solution 2: Enable keep-alive settings on your EC2 instances and set the keep-alive timeout to greater than or equal to the idle timeout settings of your load balancer.

With our architecture, this problem is exacerbated because the NLB has a non-configurable timeout of 350 seconds, and Gunicorn workers have a default of 2 seconds.

I believe this is what is resulting in 504 Gateway Timeouts.

-----

In more detail: load balancers are the midpoint between request clients (API Gateway) and the backend server. They maintain long-lived keepalive connections in order to avoid the performance costs of reconnecting on every request. Keepalive timeouts exist so that servers can close the connection after enough time has passed without activity.

If the client (NLB) has a longer timeout than the server (API), the API can close the connection without the NLB realizing. The NLB, on the next request, will try to use the connection and receive a timeout because it is already closed. This surfaces in API Gateway as a 504 (as we've seen on previous projects with ELBs)  because the API Gateway request to the NLB --> server times out.

Specifically for NLB, it also shows in metrics as "Target Reset Count" (number of times the target API reset the connection). I believe "Client reset count" may also indicate the API Gateway is keeping long connections and resetting on occasion, but at a lower frequency; but I'm not really sure TBH.

## Demo

Deployed this on test around 6pm ET. Seems to bring the target reset count to zero, although 504s and client resets are still intermittent. There aren't any LB resets so I feel moderately comfortable with the idea that our service timeout is longer than the NLB timeout.

## More Resources

https://medium.com/tenable-techblog/lessons-from-aws-nlb-timeouts-5028a8f65dda
https://www.reddit.com/r/aws/comments/cvwpp7/request_differences_between_api_gateway_and_nlb/
https://www.vertica.com/docs/9.2.x/HTML/Content/Authoring/UsingVerticaOnAWS/Configurekeepalive.htm
https://serverfault.com/questions/782022/keepalive-setting-for-gunicorn-behind-elb-without-nginx