END: Allow E2E to run in environments that have non-standard employers (#2228)
END: No Ticket
Our claim generation engine has a dependency of an employer "factory", which is responsible for pulling or generating a matching employer. The default logic is to pull the employer out of the standard "pool", which only allows us to use employees of Wayne Enterprises or Umbrella Corp. This doesn't work in the perf environment, which has totally different employers loaded. This PR updates E2E test logic to allow E2E tests to run in environments that use non-standard employers.

It also updates our .env.example file to make more sense to people coming onto the project