API-1005: Filter applications in the DB query (#2269)
https://lwd.atlassian.net/browse/API-1005

Currently, the applications_get endpoint pulls in all applications from the database and filters in memory. We're seeing 504s on this endpoint, and traces of over 60 seconds (https://one.nr/09MR2AmrkwY) with broken traces because it’s maxing out at 1000 spans (a very long list of postgres table selects).