API-324: add step function to run DOR import & FINEOS export (#1715)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-324

## 🛠 Changes

- Add an AWS Step Functions state machine to run DOR import & FINEOS export steps.
- Schedule the Step Function to start automatically using EventBridge.
- Enable in test to run a full set of steps, including DOR generate (hourly to test this out).
- Enable in prod to run only DOR import and Load employers to FINEOS (daily at 12:00 UTC (currently 07:00 Eastern)).
- In other environments, default to a step function that does nothing (daily).

## ℹ️ Context for reviewers

We need to run the following steps in order:

- DOR import
- FINEOS round-trip import (not ready yet)
- Load employers to FINEOS
- FINEOS eligibility feed export

Each of these is an ECS task, AWS Step Functions can be used to run them in series.

In the test environment, we also run DOR generate to generate fake data as the first step.

To make it possible to have different functionality per environment, the step function definition (json) is in the environment directory, not the template, e.g. `infra/api/environments/prod/step_function/dor_fineos_etl.json`. We may be able to change this in future if they become identical. Step functions and terraform don't provide good ways to parameterize which steps in the step function run per environment, so this seems simplest for now.

Some weaknesses observed while implementing:

- There isn't an easy way to start a step function at a certain step (to continue in case of failure)
- There isn't a good way to include steps only in the some environments (e.g. dor_generate in test only)
- Running a task N times concurrently is possible but requires a workaround (see `load_employers_to_fineos_init` and `load_employers_to_fineos_map`)
- It currently isn't possible to enable CloudWatch Logs using terraform https://github.com/hashicorp/terraform-provider-aws/issues/12192
- The name of a scheduled execution isn't customizable

## 👀 Demo

![Screenshot_2020-12-10_22-29-15](https://user-images.githubusercontent.com/3811269/101858251-34341280-3b37-11eb-9453-a666fe543c28.png)
![Screenshot_2020-11-06_22-27-00](https://user-images.githubusercontent.com/3811269/98431409-07e22d80-2083-11eb-86d0-6a88a690a2c8.png)

Reviewers: @apopma