API-999 Prod-Check Job Not Working as Expected (#2254)
🎫 Ticket
https://lwd.atlassian.net/browse/API-999

🛠 Changes
added prod-check as a dependency to the update-the-deploy-branch job.
ℹ️ Context for reviewers
As noted in the ticket, the query for the organization field always resolved to an empty string; causing tests to always fail. Secondly, when this test fails, the deploys continue. This PR resolves this behavior so the check can work as expected