API-998: Add employee log for new employer during DOR import. (#2284)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-998

## 🛠 Changes

- Insert employee log row for new employer for an existing employee during DOR import.

## ℹ️ Context for reviewers

Currently we only update employee row if there is a change in the employees name. Since we don't update the employee when there is a wage entry for a new employer, this change will go undetected. During the fineos eligibility export we will not generate new information for the employee with a new employer.

This PR detects this change during DOR import, and manually inserts an employee log entry to indicate the change for the eligibility export.

## 👀 Demo

`make test args=tests/dor/importer/test_import_dor.py::test_log_employees_with_new_employers`

## 🌡 End to End testing

N/A