API: Fix bug in logic of get_latest_employer_for_updates (#1495)
Update the query to enforce an ordering based on `maxdate` and some misc.
refactoring to make the behavior more contained and testable.

The code previously just picked the first appearance of the employee in the
query results, which may not be the most recent entry.

We likely need to actually export the updated employee record for every employer
on record for them. But tweaking that will follow.