API-999 Prod-Check Not Working as Expected (#2322)
🎫 Ticket
https://lwd.atlassian.net/browse/API-999

🛠 Changes
removed always() from the conditional
ℹ️ Context for reviewers
Extends this PR
in some runs, the deploy will continue even though the check failed.
Info on always()
always() was originally added to prevent the use case where the prod-check job doesn't run, subsequent jobs will get skipped. However, it seems that always() will never stop a job from running. (Which was odd, because the workflow had an &&)

👀 Demo
Example of a correct account

🌡 End to End testing
N/A