API-1021: Apply Filter To Identify Employee Update Extract (#2327)
As the employee update extract is in a folder with many other file extracts it is necessary to identify the file by name. A filter was applied to the code to get the file to process to match on the agreed on file name, without the date portion.

Test cases updated to add an extra file in the folder used to pick up the file during test case execution.