END: Allow retries on fineos claim entry by moving Fineos prep step (#2376)
Adjusts the fineos_create_claim test to call the Fineos setup steps in the test callback itself rather than in a before step. Having the beforeFineos function in a before step doesn't actually work, because on retries, before callbacks are not invoked. Going forward, we should make sure all prep steps are called inside the test callback rather than in before, to prevent retries from failing.

Note: It is actually ok to use beforeEach - that does get invoked, even on retries. The catch there is that if we have a scenario that drives to the portal, then over to fineos, beforeEach() would be invoked for each one. I think it's probably better to be consistent here and just invoke beforeFineos() and beforePortal() inside the it() block every time, even where we could use beforeEach().

* END: Allow retries on fineos claim entry by moving Fineos prep step

* Also move portal beforeEach() for consistency

* Lint fixes