END-496: Only run stable E2E tests on deployments to environments other than test (#2129)
During preparation of the hotfix, Sheldon noted that the E2E test suite is running both the "stable" and "unstable" tests during a deployment to a particular environment. The intention of splitting our test suite into stable and unstable tests was to continue capturing data for tests on "unstable" features that haven't launched yet (eg: 12/15 claim submission), while preventing those tests from failing CI. In short, only "stable" tests should considered during the release process, and running these unstable tests and potentially having them time out (or just occupying CI time during the path to production) is unnecessary and confusing.

This PR adjusts the way we trigger these tests so that when deploying to any environment other than test, the unstable tests are not run automatically. This should shorten the CI time required to release considerably, and avoid introducing noise.

Due to the way the deployment actions work (the action code that gets executed will be the code that's already on the target branch), we'd expect the next deploy to each environment to either not take advantage of this flag, or to fail (haven't figured out which, yet). I don't see a way around that.