END-338: Convert employer response and remove notification checking (#2386)
https://lwd.atlassian.net/browse/END-338

This PR converts the Employer Response test to Typescript, and removes the parts pertaining to notifications.

Note: this test also converts to using the API to submit the claim directly, rather than driving it through the portal. Should be faster, and we don't care to test the portal UI for claim submission here.