API-1041: Remove cached volume from docker-compose.release.yml (#2380)
https://lwd.atlassian.net/browse/API-1041

## 🛠 Changes

Remove cached volume from docker-compose.release.yml.

## Context

The persisted volume in docker-compose.release.yml was previously added as a workaround for the default behavior in docker-compose.yml, which mounts the entire local api directory into `/app`:

```
   volumes:
      # To make local development easier, mount project code so that we don't
      # have to rebuild the docker container every time we run
      - .:/app:z
```

This is great for regular development, but when testing the release docker build, it overwrites the compiled files that we want to use. The persisted volume was a way to "cancel out" the default mounting behavior.

However, this volume does not seem to get updated on subsequent `make release-build`s.

### Solution

Instead, remove the persisted volume and move the default mounting behavior into `docker-compose.override.yml`. This file is used by default when no `-f` argument is provided.

This has the effect of:
- Keeping the same behavior for local development (make build/make start)
- Removing the need for a workaround in docker-compose.release.yml.