END-480: Stabilize/speed up denial notice test (#2430)
https://lwd.atlassian.net/browse/END-480

Converts the denial notice test to Typescript. Adds a Cypress task to wait for documents to be added to a claim, then uses that task to wait for the denial notice to be ready before hitting the applications page in the portal. This test resolves stability and timeout issues in the test, allowing it to run in ~2 minutes.

* WIP on converting denial notice test

* More WIP code for denial notices

* Remove old test and test steps

* Reinstate before block we removed

* Restore before steps

Co-authored-by: Joseph C.A. Garcia <48695203+kingmoc@users.noreply.github.com>