API-1054: When Employer is created, the payload has a null in the OrganisationName.UpperName attribute (#2414)
Formatted values for three attributes in Organisation, OrganisationName and OrganisationDefault. The attributes are:

ShortName - Set to the first 8 characters of the legal name
UpperName - Set to the legal name with all upper case letters
UpperShortName - Set to the ShortName with all upper case letters

Added a test case to validate formatting above. Also tested via Postman to make sure adding the values allows us to search for the Organisation by the provided names.