END-526: Check for employer response e-mail as part of employer_response test (#2438)
https://lwd.atlassian.net/browse/END-526

Adds assertions for the presence and content of the employer response e-mail to the employer_response test. I'm not deeply checking the employer response e-mail because I feel that doing so is a concern of whatever component manages the templates. Our concern here is:

Was the e-mail delivered
Does it contain the link we need to act on it.

* END-526: Check for employer response e-mail as part of ...