CP-1603: Display unsupported browser banner to Internet Explorer users (#2452)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-1603

## 🛠 Changes

- Add `<UnsupportedBrowserBanner>` component for displaying to Internet Explorer users
- Add docs about Portal browser support

## ℹ️ Context for reviewers

Yes, this looks hacky, but that's the fun of targeting legacy Internet Explorer versions. ☠️  We render this as part of the exported static HTML so that the banner has no JS dependency to be rendered, helping guarantee users of unsupported browsers see the banner.

The banner message closely resembles what Mass.gov uses.