CP [infra]: Update baseline error rate alert to fill empty aggregation windows with 0% (#2454)
## 🛠 Changes

- Change baseline error rate alert policy to reduce false alarms. Fill empty windows with 0% rather than having longer windows.

## ℹ️ Context for reviewers

We previously extended the aggregation window time to 10 minutes, thinking that a longer aggregation window would improve the odds that a successful request would happen within the timeframe. That didn't pan out, and this setting seemed to increase the number of false alarms. An alternative approach that was in place over the weekend was to fill empty aggregation windows with 0%. This didn't result in any false alarms.