END: Make sure E2E test workflow has access to github token to allow reruns (#2505)
Fixes two issues with the E2E test cycle:

Rerunning the E2E test workflow in Github actions results in a no-op - it simply returns a previous run if one has happened for the specific commit. This is fixed by wiring the GITHUB_TOKEN environment variable through to Cypress.
If application submission fails, the tests hang for up to 40 seconds waiting for unstash() to error out. This PR shortens the timeout there, since we expect that file should exist before the command is called, and we do not need to wait for it.