[API-938][API-695] Send Security Scans to Pagerduty (#2468)
🎫 Ticket
https://lwd.atlassian.net/browse/API-938
https://lwd.atlassian.net/browse/API-695

🛠 Changes
Generic integration added for PD alerts
Additional step added to the API-Scheduled job to send alerts to PD
ℹ️ Context for reviewers
This ticket is to configure a final step to send alerts to PD, and to configure the bandit/safety checks in this workflow to only fail on high-severity vulnerabilities. LOW or MED severity issues should be picked up automatically by PR checks and released to stage/prod within ~7 days.