converting upload fineos doc (#2623)
This PR converts upload_fineos_doc to using typescript. It also removes the portal flow and submits the claim via the API directly with a newly registered user. This shortens this test as the portal flow is not needed for the purpose of this test. The goal of this test is to show a document uploaded via Fineos will appear for the claimant within their account.

Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>