API-896: Unit test cases for fineos_vendor_export.py and bug fix for Claim Type (#2619)
Added test cases for fineos_vendor_export.py and fixed some logic issues found thanks to the new test cases. The two issues found while adding test cases, and fixed, were:

Claim type: Have to translate from FINEOS value to the value in our lookup tables. Not sure if this is planned or we need to fix our lookup values.
Associating the Claim with the Employee. We missed associating the claim with the employee when claim was new. Also associated it with the employer, if we can properly identify it in PFML from the FINEOS employe customer number.

Previous PR for this ticket had only a happy path unit test to help debug while coding. This PR adds some more granular unit tests to improve coverage.