CP-1653: Fix intermittent leave page's number fields in Firefox and IE (#2640)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-1653

## 🛠 Changes

- Use `getAttribute` to read the `inputmode` attribute in our input change handler. 

## ℹ️ Context for reviewers

`HTMLElement.inputMode` isn't supported in Firefox or Internet Explorer, so we need to read it using `getAttribute`: https://caniuse.com/mdn-api_htmlelement_inputmode

This was only affecting the one intermittent leave page because it's the only page that included form fields that had the `inputMode` _and_ `pattern` props. The other pages use `InputHours`, which handles the string conversion itself.