CP-1681: Support new API date error format (#2703)
https://lwd.atlassian.net/browse/CP-1681

Prerequisite to #2574 — show custom date errors when receiving an error type of format instead of date.

Currently, date formatting errors are being caught by pydantic and surfaced in APM as unexpected errors. These should be caught by the OpenAPI layer.

#2574 makes that change, which has a downstream effect on what the response looks like. Notably, the error type changes to format instead of date, so the Portal needs to expect that when providing user error messages.

Made these changes by searching for date: and copy-pasting them with the new format. I'm keeping the old key for now as well, so we can be backwards-compatible.