API-1245: Get leave type from VBI_REQUESTEDABSENCE_SOM.csv (#2765)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-1245

## 🛠 Changes

Since we're having trouble getting the FINEOS exports of LeavePlan_info.csv working, this uses a different field in VBI_REQUESTEDABSENCE_SOM.csv instead. 

## ℹ️ Context for reviewers

This is based on a best guess assumption that this field will work as expected. We need someone at FINEOS to verify, but until then, we need this in order for the Integration tests with CTR + FINEOS to work this week. Otherwise, GAX will not get created.

## 👀 Demo

Snippet from VBI_REQUESTEDABSENCE_SOM which makes it _seem_ like this column (right most column in the screenshot) is acceptable.

<img width="827" alt="Screen Shot 2021-01-13 at 8 59 10 PM" src="https://user-images.githubusercontent.com/67701/104547531-a17a0e00-55e3-11eb-9f32-5c1a1f2c0129.png">

## 🌡 End to End testing

N/A