END-600: Generate employers/employees according to employer sizing probabilities (#2781)
https://lwd.atlassian.net/browse/END-600

Enhances the E2E data generation tools so that they are capable of generating employers of particular "sizes" and generating more employees for employers of larger sizes. Sizes are currently "small", "medium", and "large". This is necessary to support a data load into UAT which is supposed to contain a more realistic employer/employee distribution.