API-1178: Modify Service Agreement API (#2763)
Modified service agreement API logic to not send an empty string for a leave plan value. Instead the modified FINEOS service added an attribute - AbsenceManagement - which needs to be set to false when employer is fully exempt. In addition we have to send a flag indicating to FINEOS when to unlink existing plans from the old revision so they are not linked with the new one. We are always sending this flag as we do not necessarily know if we are just adding a plan or changing the list completely.

Test cases for fineos_actions updated to match new logic. We are still awaiting information on how, or if, start_date needs to be set.