INFRA-198: List all files in file_util.list_files & recursively copy files from FINEOS (#2847)
https://lwd.atlassian.net/browse/INFRA-198

    Add a --recursive flag to fineos-bucket-tool arguments, so we can enable recursive listing for copy_dir.
    Use a boto3 paginator for list_files so that it lists all the files in an S3 prefix, not just up to 1000 files.

ℹ️ Context for reviewers

When copying BI data extracts from FINEOS, I noticed it wasn't copying anything because all of the extracts were in nested, dated directories, e.g. 2020-12-24-11-11-11/2020-12-24-11-11-11-VBIABSENCECASE.csv. This was different than the format in TEST.

Enabling --recursive fixed it, but the copying stopped after January 3rd. There are more than 1000 files in the FINEOS bucket, so we need to list more than that.