INFRA-97: Fix inefficient pip resolution in lambda builds (#2867)
https://lwd.atlassian.net/browse/INFRA-97

    Change pip install boto3 setuptools awscli aws-sam-cli to pip install aws-sam-cli
    De-duplicate lambda build yaml now that they're all using exactly the same steps.

ℹ️ Context for reviewers

Noticed today that the lambda builds are back to being slow, after being fixed for a little while 🦢

Took a quick look while waiting for my :dobby: and realized pip install aws-sam-cli should download all of its dependencies as well, so I tried just installing that and it worked fine. In the end I guess it was kind of a silly, quick fix? 😅

pip install time improvement: 6:45-7 minutes to ~20 seconds.

While making this change, I also noticed that all the lambda build jobs are exactly the same now so I merged them together into lambda-builds and added a matrix based on the lambda folder name.