API-1176:  Test cases for vendor extract state logic (#2717)
Added more granular test cases for business logic and specific test cases for state logic

I completed adding more granular test cases for many of the functions in the module but was not able to get to the state logic. I noticed the module has moved on and other changes are being made to it so stopping here and just wanting to merge what I got done so far before code moves further along.