CP-1528: Create InputCurrency component (#2894)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-1528

## 🛠 Changes

- Adds `InputCurrency` component
- Adds component to OtherLeave sections

## ℹ️ Context for reviewers

The currency mask is unique in that it's the only mask that is not also stored in the API in its masked format. In that case, strictly speaking, is it really a mask? I considered ripping out the currency masking from `Mask.js` altogether and moving it to the `InputCurrency` component, but opted for this less drastic approach first (since the current approach follows CMS's design system).

## 👀 Demo

![cp-1528](https://user-images.githubusercontent.com/2473869/106028789-3f101b80-6081-11eb-80ca-bba96048fa9c.gif)
<img width="771" alt="Screen Shot 2021-01-27 at 8 57 54 AM" src="https://user-images.githubusercontent.com/2473869/106028826-49321a00-6081-11eb-90ea-144470b6e998.png">

## 🌡 End to End testing

Will need to check if e2e tests supply a number or string for currency value. but i don't think e2e tests exist yet for other leave section.

## 🧐 
@sawyerh