END-637: E2E configuration fixes for using perf as signoff environment (#2907)
* END-637: Update employees, employers for perf and training
Reverts the URL changes for perf introduced in #2817. Since we typically run tests from master (or from a branch based on master), this allows us to continue using this environment while the domain change is pending release. It will have to be un-reverted following deployment on Monday.
Updates the training environment in Cypress to specify an employers file. This is necessary to allow leave admin responses. We'll also have to set up the necessary leave admin accounts for this environment as well.
Blocks New Relic in Fineos. This was done before, but the way we did it wasn't working anymore.