INFRA-19 Add Additional Technical Context to FineosError Events (#2895)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-19

🛠 Changes
added fields to FineosError
E.g. fineos.request.response_time → the timing data of a failed request to Fineos
renamed a few fields to reflect where the data originated (e.g. response.status → fineos.response.status )
ℹ️ Context for reviewers
there is still some relevant data related to failed FINEOS calls that should be present in New Relic, but currently is only present in CloudWatch. This ticket is to add that additional data to the FineosError event body, which is sent to New Relic every time a call to FINEOS cannot succeed.