API-1350: Pass "notification reason" field when creating cases in FINEOS (#3009)
Added logic in fineos_actions.py to populate the notificationReason attribute in the AbsenceCase. Details of actual mapping are in the ticket. We are mapping the leave_reason value sent by the Portal to a value provided by CPS.

The notificationReason attribute is not required by the absence case FINEOS API but as part of the new regs a rule was configured in FINEOS to use notificatonReason values to route cases to different areas.