END: Add retry functionality to ER submission (#3052)
Adds retry capability to submitting employer responses. Employer responses may only be submitted after Fineos hits the API with a notification call. This only happens after a claim is submitted, leaving us with a possible race condition where we may in some cases try to submit an employer response before we're allowed to do so (because the notification call hasn't happened). This PR works around that issue by adding retry capability to the PortalSubmitter class - if we fail to fetch the data about the claim as the leave admin, we retry for up to 30 seconds. This should give the system time to resolve the race condition and allow response.

* Add retry functionality to ER submission

* Shorten retry, remove console.log