remove uneeded intercept for docs (#3090)
This is a quick removal of a cy.intercept that we were not using. This was first added as a precursor for testing the response of the PFML api document request. Based on the new routing the portal team implemented, now when a user logs in they are routed to the applications page. This results in a plethora of legal notice documents being loaded. Because we were intercepting those API calls some e2e test failed waiting for all the intercepts. We're not using this anywhere so removing this block will help get to the start application faster.
Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>