INFRA-276: Convert cases of logger.error/exception to logger.warning (#3153)
https://lwd.atlassian.net/browse/INFRA-276

🛠 Changes

    Converts some cases of logger.error/exception to logger.warning
    Converts some cases of logger.error to logger.exception to capture exc_info (stacktraces)

ℹ️ Context for reviewers

As part of integrating Sentry, we're enabling the default integration of logger.error/exception calls being captured by Sentry as errors to triage. However, there are several cases in the API that are logging at the ERROR level when they are not necessarily fatal issues. These should be logged as WARNING (with exc_info=True if they want to capture exception stacktraces.)

Relatedly, some instances of logger.error should be logging as logger.exception so that stacktraces are captured.